+++ src/dynamic-image-replacement/components/control-panel.test.tsx
