+++ src/dynamic-image-replacement/components/control-panel.tsx
