+++ src/dynamic-image-replacement/components/hello-world.test.tsx
