+++ src/dynamic-image-replacement/components/hello-world.tsx
