+++ src/dynamic-image-replacement/index.tsx
