+++ src/dynamic-image-replacement/styled.tsx
