+++ src/dynamic-image-replacement/types.ts
