+++ src/index.ts
@@ -1,6 +1,6 @@
 import { registerComponent } from 'ub-shared';
 
-import { Component } from './hello-world';
+import { Component } from './dynamic-image-replacement';
 
 /**
  * Feel free to register whatever you need to export.
